+++ android/filamat-android/src/main/cpp/MaterialBuilder.cpp
@@ -18,7 +18,7 @@
 
 #include <filamat/MaterialBuilder.h>
 
-#include <filament/EngineEnums.h>
+#include <private/filament/EngineEnums.h>
 
 using namespace filament;
 using namespace filamat;
