+++ libs/filabridge/src/SibGenerator.cpp
@@ -16,7 +16,7 @@
 
 #include "private/filament/SibGenerator.h"
 
-#include <filament/EngineEnums.h>
+#include <private/filament/EngineEnums.h>
 
 #include "private/filament/SamplerInterfaceBlock.h"
