+++ libs/filamat/src/shaders/ShaderGenerator.h
@@ -19,7 +19,7 @@
 
 #include <algorithm>
 
-#include <filament/EngineEnums.h>
+#include <private/filament/EngineEnums.h>
 #include <filament/MaterialEnums.h>
 
 #include <filamat/MaterialBuilder.h>
